[Impeller] support half precision uniforms and half precision samplers (flutter/engine#40590)
[Impeller] support half precision uniforms and half precision samplers

+++ engine/src/flutter/impeller/entity/shaders/solid_fill.frag
@@ -5,11 +5,11 @@
 #include <impeller/types.glsl>
 
 uniform FragInfo {
-  vec4 color;
+  f16vec4 color;
 }
 frag_info;
 
-out vec4 frag_color;
+out f16vec4 frag_color;
 
 void main() {
   frag_color = frag_info.color;
